Rename package names in the project
- Updated package names for improved readability and alignment with
conventions.
