+++ ChrzescijanskaRandkaQA/test-output/Default suite/Default test.html
@@ -55,11 +55,11 @@ function toggleAllBoxes() {
 <body>
 <h2 align='center'>Default test</h2><table border='1' align="center">
 <tr>
-<td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
+<td>Tests passed/Failed/Skipped:</td><td>5/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Sun Nov 19 19:19:15 CET 2023</td>
+<td>Started on:</td><td>Sun Nov 19 19:34:24 CET 2023</td>
 </tr>
-<tr><td>Total time:</td><td>31 seconds (31881 ms)</td>
+<tr><td>Total time:</td><td>88 seconds (88313 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -75,10 +75,30 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='com.chrzescijanksaRandka.qa.testScenarios.LoginTest.VerifyLoggingOutOfTheApplication()'><b>VerifyLoggingOutOfTheApplication</b><br>Test class: com.chrzescijanksaRandka.qa.testScenarios.LoginTest</td>
+<td title='com.chrzescijanksaRandka.testScenarios.LoginTest.verifyLoggingIntoTheApplicationUsingValidCredentials()'><b>verifyLoggingIntoTheApplicationUsingValidCredentials</b><br>Test class: com.chrzescijanksaRandka.testScenarios.LoginTest</td>
 <td></td>
-<td>16</td>
-<td>com.chrzescijanksaRandka.qa.testScenarios.LoginTest@192d43ce</td></tr>
+<td>0</td>
+<td>com.chrzescijanksaRandka.testScenarios.LoginTest@393671df</td></tr>
+<tr>
+<td title='com.chrzescijanksaRandka.testScenarios.LoginTest.VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials()'><b>VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials</b><br>Test class: com.chrzescijanksaRandka.testScenarios.LoginTest</td>
+<td></td>
+<td>0</td>
+<td>com.chrzescijanksaRandka.testScenarios.LoginTest@393671df</td></tr>
+<tr>
+<td title='com.chrzescijanksaRandka.testScenarios.LoginTest.verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter()'><b>verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter</b><br>Test class: com.chrzescijanksaRandka.testScenarios.LoginTest</td>
+<td></td>
+<td>0</td>
+<td>com.chrzescijanksaRandka.testScenarios.LoginTest@393671df</td></tr>
+<tr>
+<td title='com.chrzescijanksaRandka.testScenarios.LoginTest.verifyLoggingIntoTheApplicationUsingInvalidCredentials()'><b>verifyLoggingIntoTheApplicationUsingInvalidCredentials</b><br>Test class: com.chrzescijanksaRandka.testScenarios.LoginTest</td>
+<td></td>
+<td>1</td>
+<td>com.chrzescijanksaRandka.testScenarios.LoginTest@393671df</td></tr>
+<tr>
+<td title='com.chrzescijanksaRandka.testScenarios.LoginTest.VerifyLoggingOutOfTheApplication()'><b>VerifyLoggingOutOfTheApplication</b><br>Test class: com.chrzescijanksaRandka.testScenarios.LoginTest</td>
+<td></td>
+<td>12</td>
+<td>com.chrzescijanksaRandka.testScenarios.LoginTest@393671df</td></tr>
 </table><p>
 </body>
 </html>
