+++ ChrzescijanskaRandkaQA/test-output/Default suite/Default test.xml
@@ -1,17 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="4" hostname="lenovo-Komputer" failures="0" tests="1" name="Default test" time="31.881" errors="0" timestamp="2023-11-19T19:19:47 CET">
-  <testcase classname="com.chrzescijanksaRandka.qa.testScenarios.LoginTest" name="VerifyLoggingOutOfTheApplication" time="16.187"/>
-  <testcase classname="com.chrzescijanksaRandka.qa.testScenarios.LoginTest" name="verifyLoggingIntoTheApplicationUsingValidCredentials" time="0.0">
-    <ignored/>
-  </testcase> <!-- verifyLoggingIntoTheApplicationUsingValidCredentials -->
-  <testcase classname="com.chrzescijanksaRandka.qa.testScenarios.LoginTest" name="verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter" time="0.0">
-    <ignored/>
-  </testcase> <!-- verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter -->
-  <testcase classname="com.chrzescijanksaRandka.qa.testScenarios.LoginTest" name="VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials" time="0.0">
-    <ignored/>
-  </testcase> <!-- VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials -->
-  <testcase classname="com.chrzescijanksaRandka.qa.testScenarios.LoginTest" name="verifyLoggingIntoTheApplicationUsingInvalidCredentials" time="0.0">
-    <ignored/>
-  </testcase> <!-- verifyLoggingIntoTheApplicationUsingInvalidCredentials -->
+<testsuite ignored="0" hostname="lenovo-Komputer" failures="0" tests="5" name="Default test" time="88.313" errors="0" timestamp="2023-11-19T19:35:52 CET">
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="verifyLoggingIntoTheApplicationUsingValidCredentials" time="0.8"/>
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="verifyLoggingIntoTheApplicationUsingInvalidCredentials" time="1.215"/>
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter" time="0.771"/>
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials" time="0.494"/>
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="VerifyLoggingOutOfTheApplication" time="12.573"/>
 </testsuite> <!-- Default test -->
