+++ ChrzescijanskaRandkaQA/test-output/emailable-report.html
@@ -9,10 +9,14 @@
 <table>
 <tr><th>Test</th><th># Passed</th><th># Skipped</th><th># Retried</th><th># Failed</th><th>Time (ms)</th><th>Included Groups</th><th>Excluded Groups</th></tr>
 <tr><th colspan="8">Default suite</th></tr>
-<tr><td><a href="#t0">Default test</a></td><td class="num">1</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">31,881</td><td></td><td></td></tr>
+<tr><td><a href="#t0">Default test</a></td><td class="num">5</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">88,313</td><td></td><td></td></tr>
 </table>
-<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Default suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Default test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="1">com.chrzescijanksaRandka.qa.testScenarios.LoginTest</td><td><a href="#m0">VerifyLoggingOutOfTheApplication</a></td><td rowspan="1">1700417970563</td><td rowspan="1">16187</td></tr></tbody>
+<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Default suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Default test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="5">com.chrzescijanksaRandka.testScenarios.LoginTest</td><td><a href="#m0">VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials</a></td><td rowspan="1">1700418928795</td><td rowspan="1">494</td></tr><tr class="passedeven"><td><a href="#m1">VerifyLoggingOutOfTheApplication</a></td><td rowspan="1">1700418939641</td><td rowspan="1">12573</td></tr><tr class="passedeven"><td><a href="#m2">verifyLoggingIntoTheApplicationUsingInvalidCredentials</a></td><td rowspan="1">1700418894810</td><td rowspan="1">1215</td></tr><tr class="passedeven"><td><a href="#m3">verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter</a></td><td rowspan="1">1700418918908</td><td rowspan="1">771</td></tr><tr class="passedeven"><td><a href="#m4">verifyLoggingIntoTheApplicationUsingValidCredentials</a></td><td rowspan="1">1700418878478</td><td rowspan="1">800</td></tr></tbody>
 </table>
-<h2>Default test</h2><h3 id="m0">com.chrzescijanksaRandka.qa.testScenarios.LoginTest#VerifyLoggingOutOfTheApplication</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h2>Default test</h2><h3 id="m0">com.chrzescijanksaRandka.testScenarios.LoginTest#VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m1">com.chrzescijanksaRandka.testScenarios.LoginTest#VerifyLoggingOutOfTheApplication</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m2">com.chrzescijanksaRandka.testScenarios.LoginTest#verifyLoggingIntoTheApplicationUsingInvalidCredentials</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m3">com.chrzescijanksaRandka.testScenarios.LoginTest#verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m4">com.chrzescijanksaRandka.testScenarios.LoginTest#verifyLoggingIntoTheApplicationUsingValidCredentials</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 </body>
 </html>
