+++ ChrzescijanskaRandkaQA/test-output/junitreports/TEST-com.chrzescijanksaRandka.testScenarios.LoginTest.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="lenovo-Komputer" failures="0" tests="5" name="com.chrzescijanksaRandka.testScenarios.LoginTest" time="88.251" errors="0" timestamp="2023-11-19T19:35:53 CET" skipped="0">
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="verifyLoggingIntoTheApplicationUsingValidCredentials" time="15.294"/>
+  <system-out/>
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="verifyLoggingIntoTheApplicationUsingInvalidCredentials" time="11.587"/>
+  <system-out/>
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter" time="9.904"/>
+  <system-out/>
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials" time="23.371"/>
+  <system-out/>
+  <testcase classname="com.chrzescijanksaRandka.testScenarios.LoginTest" name="VerifyLoggingOutOfTheApplication" time="28.095"/>
+  <system-out/>
+</testsuite> <!-- com.chrzescijanksaRandka.testScenarios.LoginTest -->
