+++ ChrzescijanskaRandkaQA/test-output/old/Default suite/classes.html
@@ -4,7 +4,7 @@
 <th>Method name</th>
 <th>Groups</th>
 </tr><tr>
-<td>com.chrzescijanksaRandka.qa.testScenarios.LoginTest</td>
+<td>com.chrzescijanksaRandka.testScenarios.LoginTest</td>
 <td>&nbsp;</td><td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@Test</td>
