+++ ChrzescijanskaRandkaQA/test-output/old/Default suite/methods-alphabetical.html
@@ -1,10 +1,34 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="95698a">  <td>23/11/19 19:19:30</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginTest.VerifyLoggingOutOfTheApplication()[pri:5, instance:com.chrzescijanksaRandka.qa.testScenarios.LoginTest@192d43ce]">VerifyLoggingOutOfTheApplication</td> 
-  <td>main@757004314</td>   <td></td> </tr>
-<tr bgcolor="95698a">  <td>23/11/19 19:19:15</td>   <td>-14985</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;LoginTest.setup()[pri:0, instance:com.chrzescijanksaRandka.qa.testScenarios.LoginTest@192d43ce]">&gt;&gt;setup</td> 
-<td>&nbsp;</td>  <td>main@757004314</td>   <td></td> </tr>
-<tr bgcolor="95698a">  <td>23/11/19 19:19:46</td>   <td>16191</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;LoginTest.tearDown()[pri:0, instance:com.chrzescijanksaRandka.qa.testScenarios.LoginTest@192d43ce]">&lt;&lt;tearDown</td> 
-<td>&nbsp;</td>  <td>main@757004314</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:28</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginTest.VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials()[pri:4, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials</td> 
+  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:39</td>   <td>10846</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginTest.VerifyLoggingOutOfTheApplication()[pri:5, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">VerifyLoggingOutOfTheApplication</td> 
+  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:29</td>   <td>1134</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;LoginTest.setup()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&gt;&gt;setup</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:29</td>   <td>1134</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;LoginTest.setup()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&gt;&gt;setup</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:29</td>   <td>1134</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;LoginTest.setup()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&gt;&gt;setup</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:29</td>   <td>1134</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;LoginTest.setup()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&gt;&gt;setup</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:29</td>   <td>1134</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;LoginTest.setup()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&gt;&gt;setup</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:52</td>   <td>23420</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;LoginTest.tearDown()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:52</td>   <td>23420</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;LoginTest.tearDown()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:52</td>   <td>23420</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;LoginTest.tearDown()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:52</td>   <td>23420</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;LoginTest.tearDown()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:52</td>   <td>23420</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;LoginTest.tearDown()[pri:0, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:34:54</td>   <td>-33985</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginTest.verifyLoggingIntoTheApplicationUsingInvalidCredentials()[pri:2, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">verifyLoggingIntoTheApplicationUsingInvalidCredentials</td> 
+  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:35:18</td>   <td>-9887</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginTest.verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter()[pri:3, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter</td> 
+  <td>main@1613095350</td>   <td></td> </tr>
+<tr bgcolor="8e97bc">  <td>23/11/19 19:34:38</td>   <td>-50317</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginTest.verifyLoggingIntoTheApplicationUsingValidCredentials()[pri:1, instance:com.chrzescijanksaRandka.testScenarios.LoginTest@393671df]">verifyLoggingIntoTheApplicationUsingValidCredentials</td> 
+  <td>main@1613095350</td>   <td></td> </tr>
 </table>
