+++ ChrzescijanskaRandkaQA/test-output/old/Default suite/methods-not-run.html
@@ -1,6 +1,2 @@
 <h2>Methods that were not run</h2><table>
-<tr><td>com.chrzescijanksaRandka.qa.testScenarios.LoginTest.verifyLoggingIntoTheApplicationUsingValidCredentials</td></tr>
-<tr><td>com.chrzescijanksaRandka.qa.testScenarios.LoginTest.verifyLoggingIntoTheApplicationUsingKeyboardKeysTabAndEnter</td></tr>
-<tr><td>com.chrzescijanksaRandka.qa.testScenarios.LoginTest.VerifyLoggingIntoTheApplicationWithoutProvidingAnyCredentials</td></tr>
-<tr><td>com.chrzescijanksaRandka.qa.testScenarios.LoginTest.verifyLoggingIntoTheApplicationUsingInvalidCredentials</td></tr>
 </table>
